+++ platform/client-admin/src/components/generic/Loading.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { IonLoading } from '@ionic/react';
-import { ILoadingProps } from '../declarations';
+import { ILoadingProps } from '../../declarations';
 
 export const Loading: React.FC<ILoadingProps> = ({ loading }) => {
   return <IonLoading
