+++ platform/client-admin/src/components/generic/NetworkBadge.tsx
