+++ platform/client-admin/src/components/index.ts
@@ -1,8 +1,8 @@
 
 export { Volunteer } from './model/Volunteer';
-export { Empty } from './Empty';
+export { Empty } from './generic/Empty';
 export { Header } from './Header';
 export { Router } from './Router';
-export { Loading } from './Loading';
+export { Loading } from './generic/Loading';
 export { ActionsList } from './model/ActionList';
-export { NetworkBadge } from './NetworkBadge';
+export { NetworkBadge } from './generic/NetworkBadge';
