+++ platform/client-admin/src/components/model/RecipientList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { IonList } from '@ionic/react';
-import { Empty } from '../Empty';
+import { Empty } from '../generic/Empty';
 import { RecipientFieldsFragment } from '../../dataFacade';
 import { Recipient } from './Recipient';
