+++ platform/client-admin/src/pages/ViewActionPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { RouteComponentProps } from 'react-router-dom'
 import { Header } from '../components/Header';
-import { Map } from '../components/Map';
+import { Map } from '../components/generic/Map';
 import { IUpdateMatchParams } from '../declarations';
 import { useUpdateVolunteerActionMutation, useFindVolunteerActionsQuery } from '../dataFacade'
 import { AutoForm, AutoFields, ErrorsField } from 'uniforms-ionic'
